backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.4
firebase-admin>=6.5.0
google-genai>=0.2.0
python-dotenv>=1.0.0
